docs(personprovider): moving personprovider to storybook to be reusable
